bug: troubleshoot Windows builds (#28)
Fixed several bugs in the download of the vcpkg cache. Also setting
the `TEMP` environment variable to non-whitelisted values disables
functionality in PowerShell used by vcpkg. We need to keep the
default value until just before running the unit tests, at which
point the default does not work, sigh.

For the Bazel build we needed to set `BAZEL_VC` to the right location,
Bazel does not seem able to automatically detect MSVC 2019.