feat: implement std::apply
This is a part of googleapis/google-cloud-cpp#3016.

Composing GCS objects consists of multiple operations under the hood.
Different subsets of the options passed to the recursive compose
operation have to be passed on to the underlying operations.

By GCS C++ client convention, options are passed as parameter packs,
hence in the general case the filtering has to be done in compile time.

An easy way to filter these parameters is to pack them in a
`std::tuple`, process the tuple and apply it via `std::apply`.
Unfortunately, `std::apply` is only available in C++14, so this PR
reimplements it in a API-compatible way.