ci: refactor definition of docker image names
Move the define-docker-variables.sh file to a location where it can be
shared across the builds that use docker
(ci/kokoro/{docker,image,readme}) and define all the image names in this
file.

This new location is already in use by other repositories, and will make
it easier to share (or generate) build scripts for all repositories in
its new location.