docs: fix instructions for CentOS 7 (and others) (#84)
I started by porting the changes from google-cloud-cpp to fix the
CentOS-7 and openSUSE/Leap instructions. In the process I noted that
using /o as the output directory was good for a Docker image, but not
for the instructions we generate from them (as I think @mr-salty tried
to tell me), so I fixed that too. Then I regenerated INSTALL.md with the
missing `sudo` commands.