ci: reduce layers readme and install Docker images
Depending on how the server is configured the Docker images may have a
limit on the number of layers (about 120), and we are getting close to
this limit in some of our builds. Each `RUN` command creates a new
layer, so we can avoid creating too many by combining `RUN` and
`WORKDIR` commands using `&&`.