chore: add release.sh script and update release doc
This release.sh script should work for all of our repos. Once we iterate on this PR, I'll update the release docs for our other repos.

Note: if you'd like to test this script yourself (nobody has run this but me), you should be able to run it against your own forks. Like I've been testing with a command like:

$ ./release/release.sh devjgm/google-cloud-cpp-spanner
then

$ ./release/release.sh -f devjgm/google-cloud-cpp-spanner
If you want to clean up after this, you can delete the tag, branch, and the release in the UI of your fork.