fix: docker image caching (#141)
The `ci/kokoro/docker/Dockerfile.ubuntu-install` file had a bug that would cause the
image rebuild to fail: it ran with `-j <nothing>` which launches too many compiler
instances, they crash, and the build would fail.

In addition to making the build too slow, this stopped the Docker image cache from
being refreshed. Which kept the build slow. And for extra fun it caused the CheckAPI
fixes that @mr-salty made in #139 not work.