refactor: copy TracingOptions from spanner (#161)
Copy the `TracingOptions` class from `google-cloud-cpp-spanner`. I need
them in `google-cloud-cpp-pubsub`, and it seems terrible to just copy
the code. The plan is to first copy the classes here (so we can start
using them in `g-c-cpp-pubsub`) and then create aliases in
`g-c-cpp-spanner` that point to the classes in this repository.