refactor: copy ConnectionOptions from spanner
The plan is to copy the class here, and later remove it from spanner, we
need it for Cloud Pub/Sub too. This is not strictly a copy, the defaults
are different for each service, for example: the endpoint. The class is
a template, parametized by traits to set the defaults. Each client
library will instantiate this class with the right set of defaults.