cleanup: use same clang-tidy in all repositories
Use the clang-tidy distributed with Fedora (currently clang-tidy-9) in
this repository. I am planning to make similar changes in the other
repositories and bring all our `clang-tidy` builds to use the same
versions of `clang-format`, `clang-tidy`, `doxygen`, `buildifier`,
`shellcheck`, and `cmake_format`.

I also updated the clang-tidy configuration to match the newer repos
(`-pubsub` and `-spanner`), and cleaned up any errors found in the
files.

Finally, fixed a number of Doxygen comments that used `<U>` (for a
template parameter) which Doxygen thought was an unordered list or
something.