ci: configs for clang-3.8 and gcc-4.8 builds
Just more builds, these are important because some of the other projects
have them, and they break when a new release of `-common` cannot compile
with those compilers. We just had that happen with `-pubsub`. There are
some formatting changes triggered by the double braces, I chose not to
argue with clang-format (I never do).