feat: release script accepts new_version arg
This PR teaches the release.sh script how to deal with version numbers
that don't start with a leading 0. This is important since Spanner is
now 1.0.0. release.sh now accepts an optional "new_version" argument to
let the user specify the desired new version number on the command line
rather than computing it from previous tags. This is useful when we want
to jump to a non-sequential version number. For example, this would have
worked for the Spanner 1.0.0 release we did a few days ago.