fix: use a safer token to workaround libstdc++ bug
The workaround put in place in #208 does not work with g++-9.3.1 on
Fedora:31. Despite this being a "newer" version of libstdc++, it does
not support `rdrand` as a token for `std::random_device` maybe because
the CPU specific tokens are not enabled.

For now, use `/dev/urandom` which seems to be supported on all Linux
distros. On other platforms we continue to use the default constructor
for `std::random_device`.

It might be possible to refine this and detect, at run-time, if `rdrand`
is supported, but that seems like an improvement that can wait a little
bit.