fix: minimize targets in overlay build_file (#279)
We want to minimize the custom targets that we add to overlay build_files, because these are deps that our customers will also have to define in their overlay build files as well. So any targets we add should be few and important.

This PR removes all of our overlay targets except for a single one called "googleapis_system_includes", which lets us include googleapis protos using angle brackets. Yes, customers will still need to define this target, but it's a single target and a simple one. If one day we can get rid of this one too, even better.